Cherry pick CHIA-1427.

+++ chia/data_layer/download_data.py
@@ -257,6 +257,7 @@ async def insert_from_delta_file(
     data_store: DataStore,
     store_id: bytes32,
     existing_generation: int,
+    target_generation: int,
     root_hashes: List[bytes32],
     server_info: ServerInfo,
     client_foldername: Path,
@@ -265,6 +266,7 @@ async def insert_from_delta_file(
     proxy_url: str,
     downloader: Optional[PluginRemote],
     group_files_by_store: bool = False,
+    maximum_full_file_count: int = 1,
 ) -> bool:
     if group_files_by_store:
         client_foldername.joinpath(f"{store_id}").mkdir(parents=True, exist_ok=True)
@@ -318,10 +320,14 @@ async def insert_from_delta_file(
                 f"Generation: {existing_generation}. Store id: {store_id}."
             )
 
-            root = await data_store.get_tree_root(store_id=store_id)
-            with open(filename_full_tree, "wb") as writer:
-                await data_store.write_tree_to_file(root, root_hash, store_id, False, writer)
-            log.info(f"Successfully written full tree filename {filename_full_tree}.")
+            if target_generation - existing_generation <= maximum_full_file_count - 1:
+                root = await data_store.get_tree_root(store_id=store_id)
+                with open(filename_full_tree, "wb") as writer:
+                    await data_store.write_tree_to_file(root, root_hash, store_id, False, writer)
+                log.info(f"Successfully written full tree filename {filename_full_tree}.")
+            else:
+                log.info(f"Skipping full file generation for {existing_generation}")
+
             await data_store.received_correct_file(store_id, server_info)
         except Exception:
             try:
